requirements.txt
PySide6
python-registry